clean up fns
